Added JobService.startJob and implementation
Includes unit tests. Closes #67

+++ eve-tracker-web/src/main/java/squareonex/evetracker/services/StorageService.java
@@ -3,8 +3,8 @@ package squareonex.evetracker.services;
 import squareonex.evetrackerdata.model.Item;
 
 public interface StorageService{
-    void add(Item item, int amount);
-    void remove(Item item, int amount);
-    boolean isAvailable(Item item, Integer amount);
-    Integer getStorageLevel(Item item);
+    void add(Item item, long amount);
+    void remove(Item item, long amount);
+    boolean isAvailable(Item item, Long amount);
+    Long getStorageLevel(Item item);
 }
