+++ eve-tracker-web/src/main/java/squareonex/evetracker/services/StorageServiceMapImpl.java
@@ -6,10 +6,10 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class StorageServiceMapImpl implements StorageService {
-    private Map<Item, Integer> map = new HashMap<>();
+    private Map<Item, Long> map = new HashMap<>();
 
     @Override
-    public void add(Item item, int amount) {
+    public void add(Item item, long amount) {
         if (map.containsKey(item)){
             amount = map.get(item) + amount;
         }
@@ -17,20 +17,20 @@ public class StorageServiceMapImpl implements StorageService {
     }
 
     @Override
-    public void remove(Item item, int amount) {
+    public void remove(Item item, long amount) {
         if (map.containsKey(item)) {
-            Integer inStorage = map.get(item);
+            Long inStorage = map.get(item);
             map.put(item, inStorage - amount);
         }
     }
 
     @Override
-    public boolean isAvailable(Item item, Integer amount) {
+    public boolean isAvailable(Item item, Long amount) {
         return map.containsKey(item);
     }
 
     @Override
-    public Integer getStorageLevel(Item item) {
+    public Long getStorageLevel(Item item) {
         return map.get(item);
     }
 }
